Refactored commands.shooter to commands
