+++ src/org/usfirst/frc4904/robot/commands/HoodDown.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.commands.shooter;
+package org.usfirst.frc4904.robot.commands;
 
 
 public class HoodDown extends HoodSet {
