+++ src/org/usfirst/frc4904/robot/commands/HoodSet.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.commands.shooter;
+package org.usfirst.frc4904.robot.commands;
 
 
 import org.usfirst.frc4904.robot.RobotMap;
