+++ src/org/usfirst/frc4904/robot/commands/HoodUp.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.commands.shooter;
+package org.usfirst.frc4904.robot.commands;
 
 
 public class HoodUp extends HoodSet {
