+++ src/org/usfirst/frc4904/robot/subsystems/Hood.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc4904.robot.subsystems;
 
 
-import org.usfirst.frc4904.robot.commands.shooter.HoodDown;
+import org.usfirst.frc4904.robot.commands.HoodDown;
 import edu.wpi.first.wpilibj.DoubleSolenoid;
 import edu.wpi.first.wpilibj.command.Subsystem;
